Display o-comments-beta when commentsTalkReplacement flag is on
A bit hacky way to render o-comments-beta instead of o-comments when the commentsTalkReplacement flags is turned on in toggler.

Wanted to keep it simple for now and that's why we are only checking the cookies. Therefore, it will only work when the flag is turned on in toggler. For development purposes, this should be enough, but it won't work if want to turn the flag on in flags-api and expect the new UI to be available in production.