Use comments-api to get display name (aka pseudonym)
Coral Talk will be the single source of truth for display names in the future. With this PR, we use comments-api to get display names for users, which uses Coral and SUDS to find display names (precedence is given to Coral).

We also change all references to the term "pseudonym" to "display name" in the UI.

NOTE:
Longroom is still using SUDS to set display names when a user logs in for the first time and doesn't have a display name yet.

Because coral-api also uses SUDS as a source, this feature will not break with this PR.

Currently, Coral doesn't have an API to create SSO users. The only way of doing that is by using the embed script at the client side.

We need to speak to Coral team for this and revisit Longroom when it is doable.