Remove the set pseudonym routes
Users are required to set a pseudonym on registration to longroom so I
am unable to find a way that users would be able to try and access a
page and be a registered longroom user without a pseudonym.

Removing this will make moving over to coral talk a simpler process if
we have less areas to clean up.