Disable display name input when pre-populated
If a user has a commenting display name, it is pre-populated in the
display name input. Previously, a user could update their display name
in the form but we're removing this ability by disabling the input.
We're doing this because we're unable to easily create a user using
Coral's API.