Add display name to Longroom database
If a user doesn't have a commenting display name, they will be asked to
create one when they sign up for Longroom. This display name will then
be saved in the Longroom database. Our intention is, if the person is
accepted, the display name will be saved in Coral and removed from the
Longroom database. If they are rejected, they will be removed from the
Longroom database.