Create a schema for display names
As we don't have a way to create comment users via a Coral talk endpoint we need
to store them in Longroom when a user signs up to Longroom and doesn't
already have a commenting account / display name.

They will only be stored here until the user is accepted or rejected
from Longroom. If they are accepted the display name will be removed
from the longroom database once the visit a longroom post and a
commenting account is created for them.

If they are rejected from Longroom the display name row will be removed
straight away.

This is the first step which creates a table for the display names to be
stored within.