Adds an endpoint for syncing display names
This endpoint will allow display name changes to be synced with
Longroom.

This will happen if the user signs up for Longroom and doesn't already
have a display name so creates it during the sign up process. If they
then go away and do create a commenting account and choose a different
display name it will sync across to Longroom.

They helps us manage the delay / race condition of users requesting
access to Longroom and being accepted.