Don't validate or store display name if one exists
If the user already has a display name set in their commenting account
we don't need to validate it again and we also don't need to store it in
the Longroom database.

We only need to store display names that are newly set during the
Longroom sign up process. We do this because we are unable to create a
commenting account at this point and it will be created using the
provided display name once they are accepted and view a longroom post